move the buffer functions to the library
