+++ imap/message.c
@@ -36,7 +36,6 @@
 #include "account.h"
 #include "bcache.h"
 #include "body.h"
-#include "buffer.h"
 #include "context.h"
 #include "envelope.h"
 #include "globals.h"
