+++ imap/util.c
@@ -40,7 +40,6 @@
 #include "mutt.h"
 #include "account.h"
 #include "bcache.h"
-#include "buffer.h"
 #include "context.h"
 #include "globals.h"
 #include "header.h"
