+++ muttlib.c
@@ -46,7 +46,6 @@
 #include "address.h"
 #include "alias.h"
 #include "body.h"
-#include "buffer.h"
 #include "charset.h"
 #include "envelope.h"
 #include "filter.h"
@@ -58,6 +57,7 @@
 #include "list.h"
 #include "mailbox.h"
 #include "mime.h"
+#include "mutt.h"
 #include "mutt_curses.h"
 #include "mutt_regex.h"
 #include "mx.h"
