+++ pattern.c
@@ -40,7 +40,6 @@
 #include "pattern.h"
 #include "address.h"
 #include "body.h"
-#include "buffer.h"
 #include "context.h"
 #include "copy.h"
 #include "envelope.h"
@@ -52,6 +51,7 @@
 #include "lib/lib.h"
 #include "list.h"
 #include "mailbox.h"
+#include "mutt.h"
 #include "mutt_curses.h"
 #include "mutt_menu.h"
 #include "mutt_regex.h"
