+++ sort.c
@@ -27,7 +27,6 @@
 #include "sort.h"
 #include "address.h"
 #include "body.h"
-#include "buffer.h"
 #include "context.h"
 #include "envelope.h"
 #include "globals.h"
